full pickle/Barrel/Shipment testing

+++ shared/DataStructures.py
@@ -17,13 +17,6 @@ pathmap = {
     'barrel':'data/store/barrels'
     }
 
-#-------------------------------------------------------------------------------------------------------------------------------------
-# base Classes
-#-------------------------------------------------------------------------------------------------------------------------------------
-#
-
-# these are methods and attributes shared by folders and files alike that have a date_pointer
-# (though maybe that is not the right way to model?)
 
 class DatePointer():
 
@@ -62,7 +55,7 @@ class DatePointer():
                               str(self.hour),
                               self.route]))
 
-# parent class for DataLake, DataStore
+
 class GenericStore():
 
     def __init__(self, kind=None):
@@ -70,32 +63,24 @@ class GenericStore():
         self.path = self.get_path(pathmap[kind])
         self.checkpath(self.path)
 
-    # generate the unique path for this folder
     def get_path(self, prefix):
         folderpath = Path.cwd() / prefix
         self.checkpath(folderpath)
         return folderpath
 
-    # create folder if it doesn't exist, including parents
     def checkpath(self, path):
         Path(path).mkdir(parents=True, exist_ok=True)
         return
 
-    # # round a timestamp down to the hour to use as a date_pointer
-    # def timestamp_to_date_pointer(self, timestamp):
-    #     return datetime.datetime.now().replace(microsecond=0, second=0, minute=0)
-
-    # create a datepointer object by parsing a path
     def date_pointer_from_a_path(self, apath):
         parts = apath.split('/')
         return DatePointer(datetime.datetime(year=int(parts[-5]),
                                     month=int(parts[-4]),
                                     day=int(parts[-3]),
                                     hour=int(parts[-2])
-                                    )
-                           )
+                                    ))
 
-# parent class for Puddle, Barrel, ?Archive, ?Warehouse
+
 class GenericFolder():
 
     def __init__(self, date_pointer, kind=None):
@@ -103,13 +88,11 @@ class GenericFolder():
         self.path = self.get_path(pathmap[kind], date_pointer)
         self.checkpath(self.path)
 
-    # generate the unique path for this folder
     def get_path(self, prefix, date_pointer):
         folderpath = Path.cwd() / prefix / date_pointer.purepath
         self.checkpath(folderpath)
         return folderpath
 
-    # create folder if it doesn't exist, including parents
     def checkpath(self, path):
         Path(path).mkdir(parents=True, exist_ok=True)
         return
@@ -126,11 +109,7 @@ class GenericFolder():
         rm_tree(self.path)
         return
 
-#-------------------------------------------------------------------------------------------------------------------------------------
-# Puddles in a DataLake are rendered into Glaciers
-#-------------------------------------------------------------------------------------------------------------------------------------
 
-# a DataLake instance represents all of the Puddles and Glaciers
 class DataLake(GenericStore):
 
     def __init__(self):
@@ -138,38 +117,23 @@ class DataLake(GenericStore):
         self.puddles = self.load_puddles()
         # future other metadata -- array of dates and hours covered, total # of records, etc.
 
-    # dump each response to data/puddle/YYYY/MM/DD/HH/route_id/drop_2021-04-03T12:12:12.json
     def make_puddles(self, feeds, date_pointer):
-
         for route_report in feeds:
-
             for route_id,route_data in route_report.items():
-
                 route=route_id.split('_')[1]
                 puddle_date_pointer=DatePointer(date_pointer.timestamp, route)
-
                 try:
-
-                    # prepare the puddle
                     folder = Puddle(puddle_date_pointer).path
                     filename = 'drop_{}_{}.json'.format(puddle_date_pointer.route, puddle_date_pointer.timestamp).replace(' ', 'T')
                     filepath = folder / PurePath(filename)
-
-                    # parse the response
                     route_data = route_data.json()
-
-                    # write it
                     with open(filepath, 'wt', encoding="ascii") as f:
                         json.dump(route_data, f, indent=4)
-
                 except Exception as e: # no vehicle activity?
                     print (e)
                     pass
-                    # import sys
-                    # sys.exit()
         return
 
-    # builds a list of paths to puddles — e.g. hourly route folders of JSON files
     def load_puddles(self):
         files = glob('{}/*/*/*/*/*'.format(Path.cwd() / pathmap['puddle']), recursive=True)
         dirs = filter(lambda f: os.path.isdir(f), files)
@@ -180,8 +144,6 @@ class DataLake(GenericStore):
             puddles.append(Puddle(date_pointer))
         return puddles
 
-    # creates a list of all puddles that are not in current hour and can be archived
-    #bug sort list from oldest to newest, currently seems to work the other way
     def list_expired_puddles(self):
         expired_puddles = []
         for puddle in self.puddles:
@@ -189,65 +151,52 @@ class DataLake(GenericStore):
             bottom_of_hour.route=puddle.route
             if puddle.date_pointer != bottom_of_hour:
                 expired_puddles.append(puddle)
+        #bug re-sort expired_puddles list from oldest to newest
         return expired_puddles
 
-    # tar up all the files in all expired puddles (not the current hour)
     def freeze_puddles(self):
         print('firing DataLake.freeze_puddles')
         puddles_to_archive = self.list_expired_puddles()
+        if len(puddles_to_archive) == 0:
+            print('no expired puddles to freeze')
+            return
         for puddle in puddles_to_archive:
             puddle.freeze_myself_to_glacier()
-
-        # quickly scan the whole datastore and delete empty folders
         # https://gist.github.com/roddds/aff960f47d4d1dffba2235cc34cb45fb
         for dirpath, dirnames, files in os.walk( (str(Path.cwd() / pathmap['lake']))):
             if not (files or dirnames):
                 os.rmdir(dirpath)
-
         return
 
-    # todo write DataLake.load_glaciers()
-    def load_glaciers(self):
-        return
 
-# a Puddle is a folder holding raw JSON responses for a single route, single hour
 class Puddle(GenericFolder):
 
     def __init__(self, date_pointer):
         super().__init__(date_pointer, kind='puddle')
         self.date_pointer = date_pointer
-        # fail to create if date_pointer.route doesn't exist
         if date_pointer.route is False:
             raise Exception ('tried to instantiate Puddle because you called __init__ without a value in DatePointer.route')
         self.route = date_pointer.route
 
     def freeze_myself_to_glacier(self):
         print('firing Puddle.render_myself_to_archive')
-        # round up all the files this puddle
         drops_to_freeze=[x for x in self.path.glob('*.json') if x.is_file()]
-
-        # init folder and sanity check -- is there already an archive file at this location?
         try:
             outfile = Glacier(self.date_pointer, self.route)
         except OSError:
             # future write handler for partially rendered puddles(maybe a different filename, or move it to a lost+found?)
             return
-
-        # write the tarball
         with tarfile.open(outfile.filepath, "w:gz") as tar:
                 for drop in drops_to_freeze:
                     tar.add(drop, arcname=drop.name.replace(':','-')) # tar doesnt like colons
         print ('froze {} drops to Glacier at {}'.format(len(drops_to_freeze), outfile.path))
-
-        # cleanup
         self.delete_folder()
-
         return
 
-# an Glacier is a rendered Puddle, e.g. a tarball in a date_pointer/route folder
+
 class Glacier(GenericFolder):
 
-    def __init__(self, date_pointer, kind='glacier'):
+    def __init__(self, date_pointer):
         super().__init__(date_pointer, kind='glacier')
         self.date_pointer = date_pointer
         self.route = date_pointer.route
@@ -255,30 +204,21 @@ class Glacier(GenericFolder):
         if self.exist == True:
             raise OSError ('there is already an archive at {}'.format(self.filepath))
 
-
     def check_exist(self):
-        filepath = self.path / 'glacier_{}_{}-{}-{}-{}.tar.gz'.format(self.date_pointer.route,
-                                                                      self.date_pointer.year,
-                                                                      self.date_pointer.month,
-                                                                      self.date_pointer.day,
-                                                                      self.date_pointer.hour
-                                                                      )
+        filepath = self.path / 'glacier_{}_{}-{}-{}-{}.tar.gz'.\
+            format(self.date_pointer.route,self.date_pointer.year,
+                   self.date_pointer.month,self.date_pointer.day,self.date_pointer.hour)
         if filepath.is_file() is True:
             return (True, filepath)
         elif filepath.is_file() is False:
             return (False, filepath)
 
-    # todo write Glacier.get_one
-    # retrieves the appropriate file, optionally uncompresses it?
-    # e.g. data = Glacier.thaw_one(DatePointer(datetime.datetime(year=2021, month=5, day=21, hour=10), route='M15')
-    def thaw_one(self, date_pointer, **kwargs):
+    def thaw_one(self, date_pointer, **kwargs): # todo write Glacier.thaw one
+        # retrieves the appropriate file, optionally uncompresses it?
+        # e.g. data = Glacier.thaw_one(DatePointer(datetime.datetime(year=2021, month=5, day=21, hour=10), route='M15')
         return
 
-#-------------------------------------------------------------------------------------------------------------------------------------
-# Barrels in a DataStore are rendered into Shipments
-#-------------------------------------------------------------------------------------------------------------------------------------
 
-# a DataLake instance represents all of the Barrels
 class DataStore(GenericStore):
 
     def __init__(self):
@@ -286,33 +226,23 @@ class DataStore(GenericStore):
         self.barrels = self.load_barrels()
         # future other metadata -- array of dates and hours covered, total # of records, etc.
 
-    # TODO COALFACE 1 DataStore.make_barrels
-    # dump each pickle to data/barrel/YYYY/MM/DD/HH/route_id/barrel_2021-04-03T12:12:12.dat
-    def make_barrels(self, feeds, date_pointer):
 
+    def make_barrels(self, feeds, date_pointer):
         for route_report in feeds:
             for route_id,route_data in route_report.items():
                 route = route_id.split('_')[1]
                 barrel_date_pointer=DatePointer(date_pointer.timestamp, route)
-
                 pickles = []
                 try:
-
-                    # prepare the barrel
                     folder = Barrel(barrel_date_pointer).path
                     filename = 'pickle_{}_{}.dat'.format(barrel_date_pointer.route, barrel_date_pointer.timestamp).replace(' ', 'T')
                     filepath = folder / PurePath(filename)
-
-                    # parse the response
                     route_data = route_data.json()
                     for monitored_vehicle_journey in route_data['Siri']['ServiceDelivery']['VehicleMonitoringDelivery'][0]['VehicleActivity']:
                         bus = BusObservation(route, monitored_vehicle_journey)
                         pickles.append(bus)
-
-                    # write it
                     with open(filepath, "wb") as f:
                         pickle.dump(pickles, f)
-
                 except KeyError:
                     # future find a way to filter these out to reduce overhead
                     # this is almost always the result of a route that doesn't exist, so why is it in the OBA response?
@@ -324,9 +254,6 @@ class DataStore(GenericStore):
                     pass
         return
 
-
-
-    # builds a list of paths to puddles — e.g. hourly route folders of JSON files
     def load_barrels(self):
         files = glob('{}/*/*/*/*/*'.format(Path.cwd() / pathmap['barrel']), recursive=True)
         dirs = filter(lambda f: os.path.isdir(f), files)
@@ -337,8 +264,6 @@ class DataStore(GenericStore):
             barrels.append(Barrel(date_pointer))
         return barrels
 
-    # creates a list of all barrels that are not in current hour and can be archived
-    #bug sort list from oldest to newest, currently seems to work the other way
     def list_expired_barrels(self):
         expired_barrels = []
         for puddle in self.barrels:
@@ -346,80 +271,62 @@ class DataStore(GenericStore):
             bottom_of_hour.route=puddle.route
             if puddle.date_pointer != bottom_of_hour:
                 expired_barrels.append(puddle)
-        return expired_barrels
+        return expired_barrels  #bug sort list from oldest to newest
 
-
-    # load each barrel, extract the pickled buses, bundle them into a new list and then write that to a JSON container
     def render_barrels(self):
         print('firing DataStore.render_barrels')
         barrels_to_archive = self.list_expired_barrels()
+        if len(barrels_to_archive) == 0:
+            print('no expired barrels to render')
+            return
         for barrel in barrels_to_archive:
             barrel.render_myself_to_shipment()
-
-
-        # quickly scan the whole datastore and delete empty folders
         # https://gist.github.com/roddds/aff960f47d4d1dffba2235cc34cb45fb
         for dirpath, dirnames, files in os.walk( (str(Path.cwd() / pathmap['store']))):
             if not (files or dirnames):
                 os.rmdir(dirpath)
-
         return
 
-# a Barrel is a folder holding files with multple pickled BusObservation instances parsed from responses for a single route, single hour
+
 class Barrel(GenericFolder):
 
     def __init__(self, date_pointer):
         super().__init__(date_pointer, kind='barrel')
         self.date_pointer = date_pointer
-        # fail to create if date_pointer.route doesn't exist
         if date_pointer.route is False:
             raise Exception ('tried to instantiate Barrel because you called __init__ without a value in DatePointer.route')
         self.route = date_pointer.route
 
     def render_myself_to_shipment(self):
         print('firing Barrel.render_myself_to_shipment')
-
-        # init folder and sanity check -- is there already an archive file at this location?
         try:
             outfile = Shipment(self.date_pointer, self.route)
         except OSError:
             # future write handler for partially rendered puddles(maybe a different filename, or move it to a lost+found?)
             return
-
-
-        # pack the pickles
         pickles_to_render=[x for x in glob("*.dat") if x.is_file()]
         pickle_array = []
         for picklefile in pickles_to_render:
-            # print('checking {}'.format(picklefile))
             with open(picklefile, 'rb') as pickle_file:
                 barrel = pickle.load(pickle_file)
                 for p in barrel:
                     pickle_array.append(p)
                 print ('added {} to route pickle'.format(picklefile))
-
-        #iterate over pickle_array and insert into JSON
         serial_array=[]
         for p in pickle_array:
             serial_array.append(p.to_serial())
-
         json_container = dict()
         json_container['buses'] = serial_array
-
         with open(outfile.filepath, 'w') as f:
             json.dump(json_container, f)
-
         print ('wrote {} pickles to Shipment at {}'.format(len(pickle_array), outfile.path))
-
-        # cleanup
         self.delete_folder()
-
         return
-        
-# an Shipment is a rendered Barrel, e.g. a bunch of different BusObservations concatenated together
+
+
 class Shipment(GenericFolder):
 
-    def __init__(self, date_pointer, kind='shipment'):
+    def __init__(self, date_pointer):
         super().__init__(date_pointer, kind='shipment')
         self.date_pointer = date_pointer
         self.route = date_pointer.route
@@ -428,20 +335,16 @@ class Shipment(GenericFolder):
             raise OSError ('there is already a Shipment at {}'.format(self.filepath))
 
     def check_exist(self):
-        filepath = self.path / 'glacier_{}_{}-{}-{}-{}.tar.gz'.format(self.date_pointer.route,
-                                                                      self.date_pointer.year,
-                                                                      self.date_pointer.month,
-                                                                      self.date_pointer.day,
-                                                                      self.date_pointer.hour
-                                                                      )
+        filepath = self.path / 'shipment_{}_{}-{}-{}-{}.json'.\
+            format(self.date_pointer.route,self.date_pointer.year,
+                   self.date_pointer.month,self.date_pointer.day,self.date_pointer.hour)
         if filepath.is_file() is True:
             return (True, filepath)
         elif filepath.is_file() is False:
             return (False, filepath)
 
-    # todo write Shipment.get_one
-    # retrieves the appropriate file, optionally uncompresses it?
-    # e.g. data = Shipment.ship_one(DatePointer(datetime.datetime(year=2021, month=5, day=21, hour=10), route='M15')
-    def ship_one(self, date_pointer, **kwargs):
+    def ship_one(self, date_pointer, **kwargs): # todo write Shipment.get_one
+        # retrieves the appropriate file, optionally uncompresses it?
+        # e.g. data = Shipment.ship_one(DatePointer(datetime.datetime(year=2021, month=5, day=21, hour=10), route='M15')
         return
